+++ cli/src/merge_tools/diff_working_copies.rs
@@ -11,6 +11,7 @@ use jj_lib::conflicts::ConflictMarkerStyle;
 use jj_lib::fsmonitor::FsmonitorSettings;
 use jj_lib::gitignore::GitIgnoreFile;
 use jj_lib::local_working_copy::EolConversionMode;
+use jj_lib::local_working_copy::ExecChangeSetting;
 use jj_lib::local_working_copy::TreeState;
 use jj_lib::local_working_copy::TreeStateError;
 use jj_lib::local_working_copy::TreeStateSettings;
@@ -154,6 +155,7 @@ pub(crate) fn check_out_trees(
         let tree_state_settings = TreeStateSettings {
             conflict_marker_style,
             eol_conversion_mode: EolConversionMode::None,
+            exec_change_setting: ExecChangeSetting::Auto,
             fsmonitor_settings: FsmonitorSettings::None,
         };
         let mut state = TreeState::init(store.clone(), wc_path, state_dir, &tree_state_settings)?;