+++ CHANGELOG.md
@@ -14,6 +14,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### New features
 
+* Updated the executable bit representation in the local working copy to allow
+  ignoring executable bit changes on Unix. By default we try to detect the
+  filesystem's behavior, but this can be overridden manually by setting
+  `working-copy.exec-bit-change = "respect" | "ignore"`.
+
 ### Fixed bugs
 
 ## [0.36.0] - 2025-12-03